requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.23
tqdm==4.66.1
python-dotenv==0.19.2